"Add new, experimental library Structure; It doesn't look very good yet and may be discarded later."
