--- TODO.md
@@ -1,42 +1,0 @@
-##
-
-- Fractionate classes into their own folders for organization
- - Better for Tooltips, better for Version Control
-- Add collection to tooltip
-
-
-tooltip visualization:
-
-```
-- magPyLib.
-	- Collection
-	- magnet
-		- Cube
-		- Cylinder
-		- Sphere
-	- math
-		- rotatePosition
-		- randomAxis
-		- anglesFromAxis
-	- current 
-		- Line
-		- Circular
-				
-```
-
-These stay hidden:
-
-```
-- magnet
-	- _base
-
-- math
-	- _mathLibPrivate
-	
-- _fields
-	- PM_Sphere
-	- PM_Circular
-	- PM_Cube
-	- Moment_Dipole.py
-	- Current_Line.py
-	- Current_CircularLoop.py
