+++ __init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-from magPyLib._lib.classes.collection import Collection
-from magPyLib import current, magnet, math
+from magPyLib._lib._collection import Collection
+from magPyLib import math, magnet, current
 
-
+##Not showing cool tooltips!
