+++ _lib/_base/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+from magPyLib._lib._base._base import *
+
+
+
