+++ _lib/_collection/__init__.py
@@ -1,0 +1,1 @@
+from magPyLib._lib._collection._collection import Collection
