+++ magnet/__init__.py
@@ -1,3 +1,4 @@
-from magPyLib._lib.classes.magnets import Cube
-from magPyLib._lib.classes.magnets import Cylinder
-from magPyLib._lib.classes.magnets import Sphere
+from magPyLib.magnet._cube import Cube
+from magPyLib.magnet._cylinder import Cylinder
+from magPyLib.magnet._sphere import Sphere
+
