+++ magnet/_sphere.py
@@ -1,0 +1,124 @@
+from numpy import float64,isnan,array
+from magPyLib.math._mathLibPrivate import angleAxisRotation
+import sys
+from magPyLib._lib._fields.PM_Sphere import Bfield_Sphere
+from magPyLib._lib._base import HomoMag
+
+#%% THE SPHERE CLASS
+
+class Sphere(HomoMag):
+    """ 
+    This class represents a homogeneously magnetized sphere. The magnet
+    is initialized in the canonical basis (position=[0,0,0],
+    angle=0, axis=[0,0,1]) with the center at the origin. Scalar input is
+    either integer or float. Vector input format can be either list, tuple
+    or array of any data type (float, int).
+    
+    Class Initialization (only kwargs):
+    ---------------------
+    
+    mag : vec3 [mT]
+        Set magnetization vector of magnet in units of [mT].
+        
+    dim : float [mm]
+        Set diameter of the sphere in units of [mm].
+        
+    pos=[0,0,0] : vec3 [mm]
+        Set position of the center of the magnet in units of [mm].
+    
+    angle=0 : scalar [deg]
+        Set angle of orientation of magnet in units of [deg].
+    
+    axis=[0,0,1] : vec3 []
+        Set axis of orientation of the magnet.
+    
+    Class Variables:
+    ----------------
+    
+    magnetization : arr3 [mT]
+        Magnetization vector of magnet in units of [mT].
+        
+    dimension : float [mm]
+        Sphere diameter in units of [mm].
+    
+    position : arr3 [mm]
+        Position of the center of the magnet in units of [mm].
+    
+    angle : float [deg]
+        Angle of orientation of the magnet in units of [deg].
+        
+    axis : arr3 []
+        Axis of orientation of the magnet.
+    
+    Class Methods:
+    --------------
+    setPosition(newPos) : takes vec3[mm] - returns None
+        Set `newPos` as new source position.
+    
+    move(displacement) : takes vec3[mm] - return None
+        Moves source by the `displacement` argument.
+    
+    setOrientation(angle,axis) : takes float[deg],vec3[] - returns None
+        Set new source orientation (angle and axis) to argument values.
+    
+    rotate(angle,axis,CoR=[0,0,0]) : takes float[deg],vec3[],kwarg(vec3)[mm] - returns None
+        Rotate the source by `angle` about an axis parallel to `axis` running
+        through center of rotation `CoR`.
+    
+    getB(pos) : takes vec3[mm] - returns arr3[mT]
+        Gives the magnetic field generated by the source in units of [mT]
+        at the position `pos`.
+        
+    Examples:
+    ---------
+    >>> import magPyLib as magPy
+    >>> pm = magPy.magnet.Sphere(mag=[0,0,1000],dim=1)
+    >>> B = pm.getB([1,0,1])
+    >>> print(B)
+      [22.09708691  0.          7.36569564]
+    """ 
+    def __init__(self, mag=None, dim=None, pos=(0,0,0), angle=0, axis=(0,0,1)):
+
+        
+        #inherit class homoMag
+        #   - pos, Mrot, MrotInv, mag
+        #   - moveBy, rotateBy
+        HomoMag.__init__(self,pos,angle,axis,mag)
+        
+        #secure input type and check input format of dim
+        self.dimension = float(dim)
+        if self.dimension <= 0:
+            sys.exit('Bad dim<0 for sphere')
+
+    def getB(self,pos):
+        """
+        This method returns the magnetic field vector generated by the source 
+        at the argument position `pos` in units of [mT]
+        
+        Parameters:
+        ----------
+        pos : vec3 [mm]
+            Position where magnetic field should be determined.
+        
+        Returns:    
+        --------
+        magnetic field vector : arr3 [mT]
+            Magnetic field at the argument position `pos` generated by the
+            source in units of [mT].
+        """
+        #secure input type and check input format
+        p1 = array(pos, dtype=float64, copy=False)
+        
+        #relative position between mag and obs
+        posRel = p1 - self.position
+        
+        #rotate this vector into the CS of the magnet (inverse rotation)
+        p21newCm = angleAxisRotation(self.angle,-self.axis,posRel) # pylint: disable=invalid-unary-operand-type
+        
+        #the field is well known in the magnet coordinates
+        BCm = Bfield_Sphere(self.magnetization,p21newCm,self.dimension)  # obtain magnetic field in Cm
+        
+        #rotate field vector back
+        B = angleAxisRotation(self.angle,self.axis,BCm)
+        
+        return B
