+++ math/__init__.py
@@ -5,4 +5,7 @@ Created on Thu Jan 24 09:42:43 2019
 @author: ortner
 """
 
-from magPyLib._lib.mathLibPublic import randomAxis, axisFromAngles, anglesFromAxis, rotatePosition
+from magPyLib.math._randomaxis import randomAxis
+from magPyLib.math._axisfromangles import axisFromAngles
+from magPyLib.math._anglesfromaxis import anglesFromAxis
+from magPyLib.math._rotateposition import rotatePosition
